\name{JASPAR2018-package}
\alias{JASPAR2018-package}
\docType{package}
\title{
Data package for JASPAR 2018
}
\description{
Data package for JASPAR 2018. To search this databases, please use th
e package TFBSTools.
}
\details{
\tabular{ll}{
Package: \tab JASPAR2018\cr
Version: \tab 0.99.0\cr
Date: \tab 2017-09-21\cr
Depends: \tab R (>= 3.4.0), methods\cr
License: \tab GPL-2 \cr
URL: \tab http://jaspar.genereg.net/\cr
Type: \tab Package\cr
NeedsCompilation: \tab no\cr
LazyData: \tab yes\cr
}

}

\author{
Ge Tan <ge_tan@live.com>

Maintainer: Ge Tan <ge_tan@live.com>
}
\references{
  See \url{http://jaspar.genereg.net/} for more details about JASPAR.
}
\keyword{ package }
\examples{
## list the contents that are loaded into memory
ls("package:JASPAR2018")
}
